Add files via upload
Some fields may need a little explanation.

The Rating field in the Ratings table is on a scale of 1-5 with 5 being high.
The Format field is a detailed breakdown of the format. Everything other than “Hardcover” could be considered “Paperback”.
ISBN stands for International Standard Book Number and is a unique 13 digit identifier given to each edition of a book. The ISBN is represented in the barcode and is tied to the price.
The ItemID and OrderID fields are hierarchical. An order may contain multiple items.
The Staff Comment field contains summaries and reviews for some of the books.